Port to x86_64

+++ src/finger.c
@@ -120,7 +120,7 @@ void *p)
 	s->parse = 0;
 
 	tcb = open_tcp(&lsocket, &fsocket, TCP_ACTIVE, 0,
-	 fingcli_rcv, 0, f_state, 0, (int) finger);
+	 fingcli_rcv, 0, f_state, 0, (long) finger);
 
 	finger->tcb = tcb;
 	go(argc, argv, p);
