+++ src/ftpserv.c
@@ -146,7 +146,7 @@ ftpscs(struct tcb *tcb,enum tcp_state old,enum tcp_state new)
 			return;
 		}
 		ftp->control = tcb;             /* Downward link */
-		tcb->user = (int)ftp;           /* Upward link */
+		tcb->user = (long)ftp;          /* Upward link */
 
 		/* Set default data port */
 		ftp->port.address = tcb->conn.remote.address;
@@ -409,7 +409,7 @@ ftpcommand(struct ftp *ftp)
 			Xprintf(ftp->control,sending,"RETR",arg,"");
 			if (ftp->data) ftp->data->user = 0;
 			ftp->data = open_tcp(&dport,&ftp->port,TCP_ACTIVE,
-			 0,NULL,ftpdt,ftpsds,ftp->control->tos,(int)ftp);
+			 0,NULL,ftpdt,ftpsds,ftp->control->tos,(long)ftp);
 		}
 		FREE(file);
 		break;
@@ -431,7 +431,7 @@ ftpcommand(struct ftp *ftp)
 			Xprintf(ftp->control,sending,"STOR",arg,"");
 			if (ftp->data) ftp->data->user = 0;
 			ftp->data = open_tcp(&dport,&ftp->port,TCP_ACTIVE,
-			 0,ftpdr,NULL,ftpsds,ftp->control->tos,(int)ftp);
+			 0,ftpdr,NULL,ftpsds,ftp->control->tos,(long)ftp);
 		}
 		FREE(file);
 		break;
@@ -451,7 +451,7 @@ ftpcommand(struct ftp *ftp)
 			Xprintf(ftp->control,sending,"APPE",arg,"");
 			if (ftp->data) ftp->data->user = 0;
 			ftp->data = open_tcp(&dport,&ftp->port,TCP_ACTIVE,
-			 0,ftpdr,NULL,ftpsds,ftp->control->tos,(int)ftp);
+			 0,ftpdr,NULL,ftpsds,ftp->control->tos,(long)ftp);
 		}
 		FREE(file);
 		break;
@@ -478,7 +478,7 @@ ftpcommand(struct ftp *ftp)
 			Xprintf(ftp->control,sending,"LIST",file,"");
 			if (ftp->data) ftp->data->user = 0;
 			ftp->data = open_tcp(&dport,&ftp->port,TCP_ACTIVE,
-			 0,NULL,ftpdt,ftpsds,ftp->control->tos,(int)ftp);
+			 0,NULL,ftpdt,ftpsds,ftp->control->tos,(long)ftp);
 		}
 		FREE(file);
 		break;
@@ -497,7 +497,7 @@ ftpcommand(struct ftp *ftp)
 			Xprintf(ftp->control,sending,"NLST",file,"");
 			if (ftp->data) ftp->data->user = 0;
 			ftp->data = open_tcp(&dport,&ftp->port,TCP_ACTIVE,
-			 0,NULL,ftpdt,ftpsds,ftp->control->tos,(int)ftp);
+			 0,NULL,ftpdt,ftpsds,ftp->control->tos,(long)ftp);
 		}
 		FREE(file);
 		break;
