+++ src/tcp.h
@@ -203,7 +203,7 @@ struct tcb {
 	int32 mdev;             /* Mean deviation, milliseconds */
 	int32 rtt;              /* Last received RTT (for debugging) */
 
-	int user;               /* User parameter (e.g., for mapping to an
+	long user;              /* User parameter (e.g., for mapping to an
 				 * application control block
 				 */
 	int32 quench;           /* Count of incoming ICMP source quenches */
@@ -310,7 +310,7 @@ struct tcb *open_tcp(struct socket *lsocket,struct socket *fsocket,
 	void (*r_upcall)(struct tcb *tcb,int32 cnt),
 	void (*t_upcall)(struct tcb *tcb,int32 cnt),
 	void (*s_upcall)(struct tcb *tcb,enum tcp_state old,enum tcp_state new),
-	int tos,int user);
+	int tos,long user);
 int32 recv_tcp(struct tcb *tcb,struct mbuf **bpp,int32 cnt);
 void reset_all(void);
 void reset_tcp(struct tcb *tcb);
